Extract out getSubmission on table.
